Refactor schedule state handling
